Caret should respect text background color https://bugs.webkit.org/show_bug.cgi?id=117493
Reviewed by Ryosuke Niwa.

Source/WebCore:

This patch allows the caret to become visible in the black background
by getting the caret color from the color of the element containing
the text, not the parent element that has the contentEditable attribute.

Test: editing/caret/caret-color.html

* editing/FrameSelection.cpp:
(WebCore::CaretBase::paintCaret):

LayoutTests:

This tests to make sure that the caret is visible in the black background
by getting the caret color from the color of the element(span) containing
the text, not the parent element(div) that has the contentEditable
attribute.

* editing/caret/caret-color-expected.txt: Added.
* editing/caret/caret-color.html: Added.
* platform/efl/editing/caret/caret-color-expected.png: Added.
* platform/mac/TestExpectations: Skip this test case.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@152612 268f45cc-cd09-0410-ab3c-d52691b4dbfc