Simplify SVG animated type handling in the JSC bindings generator https://bugs.webkit.org/show_bug.cgi?id=118845
Reviewed by Kentaro Hara.

Simplify IsSVGAnimatedType subroutine in the bindings generator so that
we no longer need to hardcode SVG animated types. Also remove the
CanUseFastAttribute subroutine as it is equivalent to IsSVGAnimatedType.
This allows us to simplify the GetterExpression subroutine so that
we can use fastHasAttribute() unconditionally for boolean type.

No new tests, no behavior change.

* bindings/scripts/CodeGenerator.pm:
(IsSVGAnimatedType):
(GetterExpression):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@152845 268f45cc-cd09-0410-ab3c-d52691b4dbfc