String::lower() - Skip to slow path on the first failure https://bugs.webkit.org/show_bug.cgi?id=118885
Reviewed by Andreas Kling.

In the 8 bits case, we don't need to know the state of the full string before changing characters
to their lowercase variant.
Just fail immediately and start transforming characters from the point of failure.

This avoid reading the string twice when the uppercase character is not at the end of the string.

* wtf/text/StringImpl.cpp:
(WTF::StringImpl::lower):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@153007 268f45cc-cd09-0410-ab3c-d52691b4dbfc