REGRESSION(148300) GIFs not reanimated after resumeActiveDOMObjectAndAnimations https://bugs.webkit.org/show_bug.cgi?id=119217
Reviewed by Tim Horton.

Commit 148300 introduced a new better was of restarting GIF animations after
they have been suspended. Unfortunately the method was not activated when a page
is resumed but only when it is returning from background.

This fixes GIF animations that are stopped during user interaction on Qt and other
platforms using suspendActiveDOMObjectAndAnimations.

* page/Frame.cpp:
(WebCore::Frame::resumeActiveDOMObjectsAndAnimations):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@153431 268f45cc-cd09-0410-ab3c-d52691b4dbfc