[BlackBerry] Crash at absoluteClippedOverflowRect() in PluginView::calculateClipRect() https://bugs.webkit.org/show_bug.cgi?id=119401
Patch by Jacky Jiang <zhajiang@blackberry.com> on 2013-08-01.
Reviewed by George Staikos.

JIRA 462881
The renderer of the PluginView's element has been detached due to transition
to a new load. We should return an empty contentRect if there even isn't
a renderer. This is a speculative fix based on the stack trace provided.

* plugins/blackberry/PluginViewBlackBerry.cpp:
(WebCore::PluginView::calculateClipRect):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@153592 268f45cc-cd09-0410-ab3c-d52691b4dbfc