[Curl] Cookie is not set when url string is unicode. https://bugs.webkit.org/show_bug.cgi?id=119545
Patch by peavo@outlook.com <peavo@outlook.com> on 2013-08-12
Reviewed by Brent Fulgham.

When the method setCookiesFromDOM is called with an unicode url parameter, the cookie is not correctly set.
We need to convert the cookie string to 8 bit, before passing it to the Curl api function.

* platform/network/curl/CookieJarCurl.cpp:
(WebCore::setCookiesFromDOM): Convert cookie string to 8 bit, if needed.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@153946 268f45cc-cd09-0410-ab3c-d52691b4dbfc