https://bugs.webkit.org/show_bug.cgi?id=119915 REGRESSION(r154144): ASSERTION FAILED: m_history->provisionalItem() == m_requestedHistoryItem.get()
Reviewed by Darin Adler.

Tools:

The issue was that we ended up having no CFNetwork cache in the testing session due
to an incorrect cache model. There is a number of things not implemented when it
comes to dynamically changing cache model as attempted by WebKitTestRunner, but
the easiest way to get this going is to initialize it to correct value upfront.

* WebKitTestRunner/TestController.cpp: (WTR::TestController::initialize):

LayoutTests:

* platform/mac-wk2/TestExpectations: Unskipping the failing test.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@154301 268f45cc-cd09-0410-ab3c-d52691b4dbfc