[GTK] Add support for passing test directories to run-gtk-tests https://bugs.webkit.org/show_bug.cgi?id=120296
Reviewed by Philippe Normand.

* Scripts/run-gtk-tests:
(TestRunner._get_tests_from_dir): Helper function to return all
unit tests found in a given directory.
(TestRunner._get_tests): Check the given tests passed in the
command line, so that if a directory is found the tests contained
in the directory are added to the list of tests to run.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@154595 268f45cc-cd09-0410-ab3c-d52691b4dbfc