Interpolate between CSS filter() and cached images https://bugs.webkit.org/show_bug.cgi?id=120602
Reviewed by Simon Fraser.

Source/WebCore:

Interpolate from a filter() image function to a url() image reference and
the other way around, if the input image on the filter function is the same
as the referenced image on url().

* page/animation/CSSPropertyAnimation.cpp: Replaced old
    filterBlend function with blendFilter function which is more generic
    and can be reused in multiple places.
(WebCore::blendFilterOperations):
(WebCore::blendFunc):
(WebCore::blendFilter):

LayoutTests:

Add aditional test for interpolation from url to a -webkit-filter function
and the other way around. Added an additional test to make sure that different
input images don't make start and end interpolate.

* fast/filter-image/filter-image-animation.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@155240 268f45cc-cd09-0410-ab3c-d52691b4dbfc