[GTK] 32-bit builder should run the JSC tests https://bugs.webkit.org/show_bug.cgi?id=119215
Reviewed by Philippe Normand.

Rename the TestClass member of the TestFactory and BuildAndTestFactory classes
to a more descriptive name, 'LayoutTestClass'. In these two factory classes, the layout testing-specific
steps are only added if the LayoutTestClass member was not overridden to the value of None in a
derived class.

Remove the BuildAndAPITestFactory class and replace it with the BuildAndNonLayoutTestFactory class that
is derived from the BuildAndTestFactory class and overrides the LayoutTestClass member to the value of None.
This allows for builders of the BuildAndNonLayoutTest type to run all but the layout tests. This type
is currently used by the GTK 32-bit builder - there's at the moment no interest to have the builder cover
layout testing while other types of testing are still expected to be performed.

* BuildSlaveSupport/build.webkit.org-config/config.json:
* BuildSlaveSupport/build.webkit.org-config/master.cfg:
(BuildFactory.__init__):
(TestFactory):
(TestFactory.__init__):
(BuildAndTestFactory):
(BuildAndTestFactory.__init__):
(BuildAndTestWebKit2Factory):
(BuildAndTestWebKit2OnlyFactory):
(BuildAndNonLayoutTestFactory):
(TestLeaksFactory):
(TestWebKit2Factory):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@155372 268f45cc-cd09-0410-ab3c-d52691b4dbfc