Add NeverDestroyed::get https://bugs.webkit.org/show_bug.cgi?id=121358
Reviewed by Anders Carlsson.

Source/WebCore:

* html/HTMLElement.cpp:
(WebCore::HTMLElement::parseAttribute): Use auto and the new get
function to make this code less repetitive.

Source/WTF:

* wtf/NeverDestroyed.h: Removed code to make this non-copyable, since we also
use a macro to accomplish that.
(WTF::NeverDestroyed::get): Added. Useful in conjuction with auto to avoid
having to repeat the type when we want to put one of these into a reference.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@155775 268f45cc-cd09-0410-ab3c-d52691b4dbfc