[GTK] Make symbol export filter more strict, and disable for dev/test builds https://bugs.webkit.org/show_bug.cgi?id=120586
Reviewed by Martin Robinson.

.:

* GNUmakefile.am: only include test-related automake files when developer
mode is on.
* Source/autotools/ReadCommandLineArguments.m4: add --enable-developer-mode,
defaults to no, and to yes for debug builds.
* Source/autotools/SetupAutomake.m4: add ENABLE_DEVELOPER_MODE conditional.
* Source/autotools/symbols.filter: make the exported symbols list much shorter,
covering only the public ABI and a few symbols required by WebKit2 processes.

Source/WebKit/gtk:

* GNUmakefile.am: only use the version script when in developer mode..

Source/WebKit2:

* GNUmakefile.am: only use the version script if developer mode is disabled.

Tools:

* GNUmakefile.am: moved GtkLauncher and documentation builds to their own
makefiles, so they can be included unconditionally.
* GtkLauncher/GNUmakefile.am: added.
* Scripts/webkitdirs.pm:
(runAutogenForAutotoolsProjectIfNecessary): always pass --enable-developer-mode
to configure, since we consider builds done through build-webkit to be development
and test builds.
* gtk/GNUmakefile.am: added.
* gtk/common.py: do not use the Scripts directory as a reference to the top path;
it's not necessary and it will not work now that the Scripts directory is not
shipped in the tarball.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@155918 268f45cc-cd09-0410-ab3c-d52691b4dbfc