Correct range used for Emoji checks. https://bugs.webkit.org/show_bug.cgi?id=121486
Reviewed by Allan Sandfeld Jensen.

Found and reported by David Binderman via Qt bug tracker.

The check if a character was in the Emoji range always evaluated to
false due to the upper range limit being lower than the lower limit.

Changed the upper limit to the highest assigned character from the
"Transport and Map Symbols" (0x1F6C5) as that seems to have been the
intended upper range limit of this check.

* platform/graphics/Font.cpp:
(WebCore::Font::isCJKIdeographOrSymbol):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@155951 268f45cc-cd09-0410-ab3c-d52691b4dbfc