Fix accessibility-node-memory-management.html to use normal style for GC-related tests, making it more robust https://bugs.webkit.org/show_bug.cgi?id=121760
Rubber-stamped by Alexey Proskuryakov.

Change was accidentally landed in previous commit, this is the change log
that goes with it.

* accessibility/accessibility-node-memory-management.html:
Use jsTestIsAsync and setTimeout to do the garbage collection without the
rest of the script state on the stack. Also removed a lot of unneeded
"window." since the test works fine without them and tweaked comments.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156244 268f45cc-cd09-0410-ab3c-d52691b4dbfc