Do not reenter Document from its destructor https://bugs.webkit.org/show_bug.cgi?id=121933
Reviewed by Darin Adler.

Following r156422, there are cases where we would reenter Document
during its destruction.

The flow was the following:
-Document's destructor is executed.
-DocumentStyleSheetCollection being part of Document, its destructor
 is executed.
-DocumentStyleSheetCollection's destructor detach the owner node of all
 its stylesheets avoid keeping stale references to a dead Document.
-CSSStyleSheet::clearOwnerNode() would fetch the dying document
 and invoke Document::styleResolverChanged().
This would not fail because the document is detached and Document::styleResolverChanged()
would fail early. This is complicated and error prone so this patch attempt to clear
the process a bit.

With the patch, Document detach all the stylesheets early in the destruction.
Clearing the stylesheet is also changed to avoid re-entering the document.

* css/CSSStyleSheet.h:
* dom/Document.cpp:
(WebCore::Document::~Document):
* dom/DocumentStyleSheetCollection.cpp:
(WebCore::DocumentStyleSheetCollection::detachFromDocument):
* dom/DocumentStyleSheetCollection.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156444 268f45cc-cd09-0410-ab3c-d52691b4dbfc