[CSSRegions] Unable to collect html element in a named flow https://bugs.webkit.org/show_bug.cgi?id=121878
Reviewed by Antti Koivisto.

Source/WebCore:

Tests: fast/regions/collect-html-in-named-flow.html
       fast/regions/iframe-html-collected.html

It should be possible to collect the html element in a named flow.
However, before we check whether adding a child is not restricted
by its original parent, we have to make sure that the original parent
is an element to safely perform the check.

* rendering/RenderNamedFlowThread.cpp:
(WebCore::RenderNamedFlowThread::isChildAllowed):

LayoutTests:

Use an iframe with a html element collected in a named flow
to test that is does not crash or assert.

* fast/regions/collect-html-in-named-flow-expected.txt: Added.
* fast/regions/collect-html-in-named-flow.html: Added.
* fast/regions/iframe-html-collected-expected.txt: Added.
* fast/regions/iframe-html-collected.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156450 268f45cc-cd09-0410-ab3c-d52691b4dbfc