[GTK][WK2] Logging of accessibility events is not disabled between tests https://bugs.webkit.org/show_bug.cgi?id=121957
Patch by Denis Nomiyama <d.nomiyama@samsung.com> on 2013-09-27
Reviewed by Mario Sanchez Prada.

Tools:

Fixed an issue in AccessibilityNotificationHandler where the destructor
was not disconnecting accessibility event callbacks if logging was on.
The decision of disconnecting events can solely rely on the existence of
notification handlers because logging uses the global notification
handler.

* WebKitTestRunner/InjectedBundle/atk/AccessibilityNotificationHandlerAtk.cpp:
(WTR::AccessibilityNotificationHandler::disconnectAccessibilityCallbacks): Fixed
disconnection when logging is on.

LayoutTests:

Removed some accessibility tests from the expectations file that were
failing before this fix.

* platform/gtk-wk2/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156541 268f45cc-cd09-0410-ab3c-d52691b4dbfc