user-select: none cursor turns to I-beam on mouse dragging https://bugs.webkit.org/show_bug.cgi?id=90159
Patch by Vani Hegde <vani.hegde@samsung.com> on 2013-09-30
Reviewed by Ryosuke Niwa.

Source/WebCore:

When tried to select some text in an area that has style
cursor:default set, cursor type changes to text cursor ignoring
the cursor style that is explicitly set.

When the cursor style is explicitly set as default (or something else),
we should not change it to text cursor no matter what we are over
or what operation we are performing (be it hovering over the text
or selecting the text).

During text selection, changing the cursor type to text only when
there is no explicit cursor type set fixes the issue.

Test: editing/caret/selection-with-caret-type-progress.html

* page/EventHandler.cpp:
(WebCore::EventHandler::selectCursor):
During selection, set the cursor style to text only if there is no
explicit cursor style set.

LayoutTests:

Test case added to verify that explicitly set cursor style
is retained on performing text selection.

* editing/caret/selection-with-caret-type-progress-expected.txt: Added.
* editing/caret/selection-with-caret-type-progress.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156635 268f45cc-cd09-0410-ab3c-d52691b4dbfc