[InexedDB] Interfaces inheriting from EventTarget should generate JSC (un)wrapping functions https://bugs.webkit.org/show_bug.cgi?id=122114
Reviewed by Darin Adler.

After r156590, four IndexedDB interfaces are inheriting from the EventTarget interface. Because of that,
they're now required to have interface-specific JSC wrapping and unwrapping functions generated since the
JSC wrapping and unwrapping functions for the EventTarget interface delegate these operations to these
interface-specific functions. If the functions are not generated and accessible, the EventTarget functions
end up recursively calling themselves.

* Modules/indexeddb/IDBDatabase.idl:
* Modules/indexeddb/IDBOpenDBRequest.idl:
* Modules/indexeddb/IDBRequest.idl:
* Modules/indexeddb/IDBTransaction.idl:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156701 268f45cc-cd09-0410-ab3c-d52691b4dbfc