[css-shapes] shape-outside does not properly handle the container and the float having different writing modes https://bugs.webkit.org/show_bug.cgi?id=122243
Reviewed by Alexandru Chiculita.

Source/WebCore:

The margin box offset needs to be computed based on the writing mode
of the container, not the writihg mode of the float itself. This
patch makes that happen.

Test: fast/shapes/shape-outside-floats/shape-outside-floats-different-writing-modes.html

* rendering/shapes/ShapeOutsideInfo.cpp:
(WebCore::ShapeOutsideInfo::updateDeltasForContainingBlockLine):

LayoutTests:

Test that the margins and width are properly computed in the case of
different writing modes.

* fast/shapes/shape-outside-floats/shape-outside-floats-different-writing-modes-expected.html: Added.
* fast/shapes/shape-outside-floats/shape-outside-floats-different-writing-modes.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156806 268f45cc-cd09-0410-ab3c-d52691b4dbfc