Allow skipping of run-jsc-stress-tests that are configured using yaml files https://bugs.webkit.org/show_bug.cgi?id=122385
Tools: 

Reviewed by Sam Wieinig.
        
Make tests run via YAML be able to do the parse-for-commands trick that allows
us to skip tests and do other custom per-test stuff.

* Scripts/jsc-stress-test-helpers/layout-test-helper:
* Scripts/run-jsc-stress-tests:

LayoutTests: 

Reviewed by Sam Weinig.
        
The way that we "skip" tests is by having a custom command at the top using //@ notation,
and then use "//@ skip". But the machinery to do that also allows us to do things like
run tests in a no-concurrent-JIT mode - so I've also implemented that in this patch.
        
Use the skip machinery to skip large-expressions.js.

* js/script-tests/large-expressions.js:
* jsc-layout-tests.yaml:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156976 268f45cc-cd09-0410-ab3c-d52691b4dbfc