[GTK] Fails to build with jhbuild with newer system glib https://bugs.webkit.org/show_bug.cgi?id=122132
Patch by Gustavo Noronha Silva <gustavo.noronha@collabora.com> on 2013-10-07
Reviewed by Martin Robinson.

Newer glib includes new symbols that some system libraries which we depend on, such as pango, have
started relying on. glib is not a dependency that should change test results, and is provided in
jhbuild with the intention of supporting the other bits that do, so using a newer stable release
should not be an issue.

* gtk/jhbuild.modules:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@157037 268f45cc-cd09-0410-ab3c-d52691b4dbfc