[Win] Unreviewed build fix after r157228
The changes in r157228 trigger a bug in VS2010. This patch works around the compiler bug
to allow this platform to build.

* dom/Range.cpp: Move implementation to header file.
* dom/Range.h: Inline the method.
* page/DOMWindow.cpp:
(WebCore::DOMWindow::open): Add stub method (with pointer arguments) to call the real
method with references.
* page/DOMWindow.h: Declare stub method taking pointer arguments to avoid compiler bug.
* testing/Internals.cpp:
(WebCore::Internals::openDummyInspectorFrontend): Use work-around method when building under
the buggy version of Visual Studio.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@157252 268f45cc-cd09-0410-ab3c-d52691b4dbfc