FTL shouldn't pass i1's into llvm.webkit.stackmap's https://bugs.webkit.org/show_bug.cgi?id=122629 <rdar://problem/15203037>
Reviewed by Sam Weinig and Nadav Rotem.
        
LLVM's stackmap support requires that we only pass operands with legal types (i.e. types
that are hardware-representable). i1, which the FTL previously used for Booleans, is not
legal.
        
We have two options: either add support in LLVM to legalize stackmap operands, or add
support to the FTL to legalize stackmap operands. It's easier to fix this in FTL, and
that's what this patch does.

* ftl/FTLLowerDFGToLLVM.cpp:
(JSC::FTL::LowerDFGToLLVM::addExitArgumentForNode):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@157293 268f45cc-cd09-0410-ab3c-d52691b4dbfc