Make RenderLayer not arena-allocated. <https://webkit.org/b/87523>
Reviewed by Antti Koivisto.

Let RenderLayerModelObject store its RenderLayer in a unique_ptr.
Layers are relatively low-volume objects and this looks neutral
on our performance tests.

Re-landing this since the original commit exposed a problem with
the destruction order. This pattern:

    m_someRenderer->destroy();
    m_someRenderer = nullptr;

..is different from what we get with unique_ptr, which clears
the pointer storage before calling the destructor.
I believe that the new order is strictly better as it removes
one path to accessing an object that is undergoing destruction.

The exposed bug was in RLC::fixedRootBackgroundLayerChanged()
where we were dereferencing the RenderView's layer() without
checking, despite it being called below ~RenderLayer.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@157333 268f45cc-cd09-0410-ab3c-d52691b4dbfc